refactor: ensure environments are corrected in the correct place
